# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Abroad AI Counselor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiError {
  error: string;
}

export interface University {
  id: number;
  name: string;
  country: string;
  city: string;
  website?: string;
  description?: string;
  ranking?: number;
}

export interface Program {
  id: number;
  universityId: number;
  universityName: string;
  name: string;
  degree: string;
  duration: string;
  language: string;
  tuitionPerYear?: string;
  applicationDeadline?: string;
  requirements?: string;
  ieltsMin?: number;
  toeflMin?: number;
}

export interface UniversityWithPrograms {
  id: number;
  name: string;
  country: string;
  city: string;
  website?: string;
  description?: string;
  ranking?: number;
  programs: Program[];
}

export interface VisaRequirement {
  id: number;
  country: string;
  visaType: string;
  requirements: string;
  processingTime?: string;
  fees?: string;
  proofOfFunds?: string;
  notes?: string;
}

export type ListUniversitiesParams = {
  country?: string;
  search?: string;
};

export type ListProgramsParams = {
  universityId?: number;
  degree?: string;
};

export type ListVisaRequirementsParams = {
  country?: string;
  visaType?: string;
};