# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Abroad AI Counselor API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all universities in the knowledge base
 */
export const ListUniversitiesQueryParams = zod.object({
  country: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
});

export const ListUniversitiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  country: zod.string(),
  city: zod.string(),
  website: zod.string().optional(),
  description: zod.string().optional(),
  ranking: zod.number().optional(),
});
export const ListUniversitiesResponse = zod.array(ListUniversitiesResponseItem);

/**
 * @summary Get a university with its programs
 */
export const GetUniversityParams = zod.object({
  id: zod.coerce.number(),
});

export const GetUniversityResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  country: zod.string(),
  city: zod.string(),
  website: zod.string().optional(),
  description: zod.string().optional(),
  ranking: zod.number().optional(),
  programs: zod.array(
    zod.object({
      id: zod.number(),
      universityId: zod.number(),
      universityName: zod.string(),
      name: zod.string(),
      degree: zod.string(),
      duration: zod.string(),
      language: zod.string(),
      tuitionPerYear: zod.string().optional(),
      applicationDeadline: zod.string().optional(),
      requirements: zod.string().optional(),
      ieltsMin: zod.number().optional(),
      toeflMin: zod.number().optional(),
    }),
  ),
});

/**
 * @summary List all programs in the knowledge base
 */
export const ListProgramsQueryParams = zod.object({
  universityId: zod.coerce.number().optional(),
  degree: zod.coerce.string().optional(),
});

export const ListProgramsResponseItem = zod.object({
  id: zod.number(),
  universityId: zod.number(),
  universityName: zod.string(),
  name: zod.string(),
  degree: zod.string(),
  duration: zod.string(),
  language: zod.string(),
  tuitionPerYear: zod.string().optional(),
  applicationDeadline: zod.string().optional(),
  requirements: zod.string().optional(),
  ieltsMin: zod.number().optional(),
  toeflMin: zod.number().optional(),
});
export const ListProgramsResponse = zod.array(ListProgramsResponseItem);

/**
 * @summary List visa requirements by country
 */
export const ListVisaRequirementsQueryParams = zod.object({
  country: zod.coerce.string().optional(),
  visaType: zod.coerce.string().optional(),
});

export const ListVisaRequirementsResponseItem = zod.object({
  id: zod.number(),
  country: zod.string(),
  visaType: zod.string(),
  requirements: zod.string(),
  processingTime: zod.string().optional(),
  fees: zod.string().optional(),
  proofOfFunds: zod.string().optional(),
  notes: zod.string().optional(),
});
export const ListVisaRequirementsResponse = zod.array(
  ListVisaRequirementsResponseItem,
);