# lib/api-zod/src/generated/types/program.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Abroad AI Counselor API
 * OpenAPI spec version: 0.1.0
 */

export interface Program {
  id: number;
  universityId: number;
  universityName: string;
  name: string;
  degree: string;
  duration: string;
  language: string;
  tuitionPerYear?: string;
  applicationDeadline?: string;
  requirements?: string;
  ieltsMin?: number;
  toeflMin?: number;
}