# lib/api-zod/src/generated/types/universityWithPrograms.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Abroad AI Counselor API
 * OpenAPI spec version: 0.1.0
 */
import type { Program } from "./program";

export interface UniversityWithPrograms {
  id: number;
  name: string;
  country: string;
  city: string;
  website?: string;
  description?: string;
  ranking?: number;
  programs: Program[];
}