# lib/api-zod/src/generated/types/visaRequirement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Abroad AI Counselor API
 * OpenAPI spec version: 0.1.0
 */

export interface VisaRequirement {
  id: number;
  country: string;
  visaType: string;
  requirements: string;
  processingTime?: string;
  fees?: string;
  proofOfFunds?: string;
  notes?: string;
}